Move PipelineContext to its own file.
